archive sync'd mail items
This provides a way to take a given source mail item and archive
it on the source.  If an `archive_folder` in configured, that folder
will be created on the source (if needed) and the mail item copied
to it, marked as seen, and deleted from the inbox.  If no
`archive_folder` is specified, messages will be deleted from the
inbox with no archive.