send notifications when errors occur
The idea is that the syncs will be running in some bg daemon or
something.  When an error occurs, the sync needs to send a notification
that an error has occured the user knows they aren't receiving some
email.

The errors are still logged but a message will be sent using the
`notify` config settings.  If an error occurs sending the notice
(ie Timeout::Error), it is simply logged and the sync continues
running.