runner interrupt
This breaks up the runner looping algorithm and allows for runner
interruption.

Signals are trapped and instead of editing some shared state to
communicate a signaled shutdown, it instead raises an `Interrupt`
exception.  The interrupt is then rescued, both in the main thread
and while waiting for the sync thread to finish.  If an Interrupt
is rescued, a stop is signaled and the main loop continues waiting
for the sync thread to join in.

This ensures that the runner will always wait for the sync thread
to join but means that a shutdown can be signaled while the main
thread is sleeping.

The sync checks if the runner has been signaled to shutdown while
processing mail items, so the sync thread should at most finish
its current mail item before aborting and joining the main thread.