lazy fetch mail item attrs
Wait to fetch mail attrs for a mail item until they are used.  This
speeds up the performance of the `MailItem.find` method significantly
as the uids are the only pieces of data fetched at find time.

Each mail item now tracks its uid and imap object.  When a mail item
attribute is used for the first time, it will fetch its attrs using
its imap and uid and cache them for use.